separated proposal class
